+++ concordium-smart-contract-testing/src/types.rs
@@ -110,7 +110,7 @@ pub struct ModuleDeploySuccess {
     pub transaction_fee:  Amount,
 }
 
-/// An error that occured while deploying a [`ContractModule`].
+/// An error that occurred while deploying a [`ContractModule`].
 #[derive(Debug, Error)]
 #[error(
     "Module deployment failed after consuming {energy_used}NRG ({transaction_fee} microCCD) with \
@@ -126,7 +126,7 @@ pub struct ModuleDeployError {
     pub kind:            ModuleDeployErrorKind,
 }
 
-/// The specific kind of error that occured while deploying a
+/// The specific kind of error that occurred while deploying a
 /// [`ContractModule`].
 #[derive(Debug, Error)]
 pub enum ModuleDeployErrorKind {
@@ -196,7 +196,7 @@ pub struct ContractInitSuccess {
     pub transaction_fee:  Amount,
 }
 
-/// An error that occured in [`Chain::contract_init`].
+/// An error that occurred in [`Chain::contract_init`].
 #[derive(Debug, Error)]
 #[error(
     "Contract initialization failed after consuming {energy_used}NRG ({transaction_fee} microCCD) \
@@ -265,7 +265,7 @@ pub enum InitExecutionError {
     OutOfEnergy,
 }
 
-/// An error that occured while executing a contract init or receive function.
+/// An error that occurred while executing a contract init or receive function.
 #[derive(Debug, Error)]
 #[error("The contract execution halted due to: {0}")]
 pub struct ExecutionError(#[from] pub(crate) anyhow::Error);
@@ -273,9 +273,9 @@ pub struct ExecutionError(#[from] pub(crate) anyhow::Error);
 /// Represents a successful contract update (or invocation).
 #[derive(Debug)]
 pub struct ContractInvokeSuccess {
-    /// Host events that occured. This includes interrupts, resumes, and
+    /// Host events that occurred. This includes interrupts, resumes, and
     /// upgrades.
-    pub trace_elements:  Vec<ContractTraceElement>,
+    pub trace_elements:  Vec<DebugTraceElement>,
     /// Energy used.
     pub energy_used:     Energy,
     /// Cost of transaction.
@@ -292,8 +292,11 @@ impl ContractInvokeSuccess {
     /// Extract all the events logged by all the contracts in the invocation.
     /// The events are returned in the order that they are emitted, and are
     /// paired with the address of the contract that emitted it.
+    ///
+    /// Only events from effective trace elements are included. See
+    /// [`Self::effective_trace_elements`] for more details.
     pub fn events(&self) -> impl Iterator<Item = (ContractAddress, &[ContractEvent])> {
-        self.trace_elements.iter().flat_map(|cte| {
+        self.effective_trace_elements().flat_map(|cte| {
             if let ContractTraceElement::Updated { data } = cte {
                 Some((data.address, data.events.as_slice()))
             } else {
@@ -306,10 +309,13 @@ impl ContractInvokeSuccess {
     /// invocation. The return value is an iterator over triples `(from, amount,
     /// to)` where `from` is the sender contract, and `to` is the receiver
     /// account. The transfers are returned in the order that they occurred.
+    ///
+    /// Only tranfers from effective trace elements are included. See
+    /// [`Self::effective_trace_elements`] for more details.
     pub fn account_transfers(
         &self,
     ) -> impl Iterator<Item = (ContractAddress, Amount, AccountAddress)> + '_ {
-        self.trace_elements.iter().flat_map(|cte| {
+        self.effective_trace_elements().flat_map(|cte| {
             if let ContractTraceElement::Transferred { from, amount, to } = cte {
                 Some((*from, *amount, *to))
             } else {
@@ -318,11 +324,44 @@ impl ContractInvokeSuccess {
         })
     }
 
-    /// Get the trace elements grouped by which contract they originated
-    /// from.
+    /// Get an iterator over references of all the [`ContractTraceElement`]s
+    /// that have *not* been rolled back.
+    ///
+    /// The trace elements returned here corresponds to the ones returned by the
+    /// node.
+    ///
+    /// See also [`Self::effective_trace_elements_cloned`] for a version with
+    /// clones.
+    pub fn effective_trace_elements(&self) -> impl Iterator<Item = &ContractTraceElement> {
+        self.trace_elements.iter().filter_map(|cte| match cte {
+            DebugTraceElement::Regular { trace_element, .. } => Some(trace_element),
+            DebugTraceElement::WithFailures { .. } => None,
+        })
+    }
+
+    /// Get an iterator over clones of all the [`ContractTraceElement`]s that
+    /// have *not* been rolled back.
+    ///
+    /// The trace elements returned here corresponds to the ones returned by the
+    /// node.
+    ///
+    /// See also [`Self::effective_trace_elements`] for a version with
+    /// references.
+    pub fn effective_trace_elements_cloned(&self) -> Vec<ContractTraceElement> {
+        self.trace_elements
+            .iter()
+            .filter_map(|cte| match cte {
+                DebugTraceElement::Regular { trace_element, .. } => Some(trace_element.clone()),
+                DebugTraceElement::WithFailures { .. } => None,
+            })
+            .collect()
+    }
+
+    /// Get the successful trace elements grouped by which contract they
+    /// originated from.
     pub fn trace_elements(&self) -> BTreeMap<ContractAddress, Vec<ContractTraceElement>> {
         let mut map: BTreeMap<ContractAddress, Vec<ContractTraceElement>> = BTreeMap::new();
-        for event in self.trace_elements.iter() {
+        for event in self.effective_trace_elements() {
             map.entry(Self::extract_contract_address(event))
                 .and_modify(|v| v.push(event.clone()))
                 .or_insert_with(|| vec![event.clone()]);
@@ -345,12 +384,12 @@ impl ContractInvokeSuccess {
         }
     }
 
-    /// Get the contract updates that happened in the transaction.
+    /// Get the successful contract updates that happened in the transaction.
     /// The order is the order of returns. Concretely, if A calls B (and no
     /// other calls are made) then first there will be "B updated" event,
     /// followed by "A updated", assuming the invocation of both succeeded.
     pub fn updates(&self) -> impl Iterator<Item = &InstanceUpdatedEvent> {
-        self.trace_elements.iter().filter_map(|e| {
+        self.effective_trace_elements().filter_map(|e| {
             if let ContractTraceElement::Updated { data } = e {
                 Some(data)
             } else {
@@ -358,9 +397,75 @@ impl ContractInvokeSuccess {
             }
         })
     }
+
+    /// Check whether any rollbacks occurred.
+    ///
+    /// That is, whether any contract calls failed which lead to state and
+    /// balances being rolled back.
+    ///
+    /// If `true` is returned, the relevant traces can be seen in the
+    /// `self.trace_elements` vector.
+    pub fn rollbacks_occurred(&self) -> bool {
+        self.trace_elements
+            .iter()
+            .any(|element| matches!(element, DebugTraceElement::WithFailures { .. }))
+    }
 }
 
-/// An error that occured during a [`Chain::contract_update`] or
+/// A wrapper for [`ContractTraceElement`], which provides additional
+/// information for testing and debugging. Most notably, it contains trace
+/// elements for failures, which are normally discarded by the node.
+#[derive(Debug)]
+pub enum DebugTraceElement {
+    /// A regular trace element with some additional data, e.g., energy usage
+    /// and the entrypoint.
+    /// This variant may be included included in the `WithFailures` list of
+    /// trace elements.
+    Regular {
+        /// The entrypoint.
+        entrypoint:    OwnedEntrypointName,
+        /// The trace element.
+        trace_element: ContractTraceElement,
+        /// The energy used so far.
+        energy_used:   Energy,
+    },
+    /// One or multiple trace elements that fail. Useful for debugging.
+    /// This variant also contains additional information, such as the error,
+    /// entrypoint, and energy usage.
+    WithFailures {
+        /// The address of the contract which failed.
+        /// This will always match the address in the last element in
+        /// `trace_elements` if the vector isn't empty.
+        contract_address: ContractAddress,
+        /// The entrypoint which failed.
+        /// This will always match the address in the last element in
+        /// `trace_elements` if the vector isn't empty.
+        entrypoint:       OwnedEntrypointName,
+        /// The error returned.
+        error:            InvokeExecutionError,
+        /// Intermediate [`DebugTraceElement`]s which occurred prior to failing.
+        /// These are the elements which are normally discared by the node.
+        trace_elements:   Vec<DebugTraceElement>,
+        /// The energy used so far.
+        energy_used:      Energy,
+    },
+}
+
+/// The reason for why a contract invocation failed during execution.
+#[derive(Debug)]
+pub enum InvokeExecutionError {
+    /// The contract rejected.
+    Reject {
+        /// The error code for why it rejected.
+        reason:       i32,
+        /// The return value.
+        return_value: ReturnValue,
+    },
+    /// The contract trapped.
+    Trap { error: ExecutionError },
+}
+
+/// An error that occurred during a [`Chain::contract_update`] or
 /// [`Chain::contract_invoke`].
 #[derive(Debug, Error)]
 #[error(
@@ -373,6 +478,8 @@ pub struct ContractInvokeError {
     /// The transaction fee. For [`Chain::contract_update`], this is the amount
     /// charged to the `invoker` account.
     pub transaction_fee: Amount,
+    /// Trace elements that occurred before the contract failed.
+    pub trace_elements:  Vec<DebugTraceElement>,
     /// The specific reason for why the invocation failed.
     pub kind:            ContractInvokeErrorKind,
 }