+++ concordium-smart-contract-testing/tests/checkpointing.rs
@@ -78,7 +78,7 @@ fn test_case_1() {
         Amount::zero(),
     );
 
-    chain
+    let update = chain
         .contract_update(
             Signer::with_one_key(),
             ACC_0,
@@ -95,6 +95,34 @@ fn test_case_1() {
             },
         )
         .expect("Updating contract should succeed");
+
+    // Check that rollbacks occurred.
+    assert!(update.rollbacks_occurred());
+
+    // Check that all the trace elements are as expected, including the ones
+    // resulting in a failure. Some imports to simplify the names in the assert.
+    use ContractTraceElement::*;
+    use DebugTraceElement::*;
+    use InvokeExecutionError::*;
+    assert!(matches!(&update.trace_elements[..], [
+        Regular {
+            trace_element: Interrupted { .. },
+            ..
+        },
+        WithFailures {
+            error: Trap { .. },
+            trace_elements,
+            ..
+        },
+        Regular {
+            trace_element: Resumed { .. },
+            ..
+        },
+        Regular {
+            trace_element: Updated { .. },
+            ..
+        }
+    ] if matches!(trace_elements[..], [Regular { trace_element: Interrupted {..}, ..}, Regular { trace_element: Updated {..}, .. }, Regular { trace_element: Resumed {..}, .. }])));
 }
 
 /// This test has the following call pattern:
@@ -202,6 +230,9 @@ fn test_case_2() {
     assert_eq!(update_a_modify_proxy.address, res_init_a.contract_address);
     assert_eq!(update_a_modify_proxy.receive_name, "a.a_modify_proxy");
     assert!(updates.next().is_none(), "No more updates expected.");
+
+    // Check that no rollbacks occurred.
+    assert!(!trace.rollbacks_occurred());
 }
 
 /// This test has the following call pattern:
@@ -345,7 +376,7 @@ fn test_case_4() {
         Amount::zero(),
     );
 
-    chain
+    let update = chain
         .contract_update(
             Signer::with_one_key(),
             ACC_0,
@@ -364,4 +395,7 @@ fn test_case_4() {
             },
         )
         .expect("Updating contract should succeed");
+
+    // Check that no rollbacks occurred.
+    assert!(!update.rollbacks_occurred());
 }